+++ .github/workflows/composite/unit-integration-test-runner/run-unit-tests.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+set -e
+
+if [ -d "/usr/lib/x86_64-linux-gnu/" ]; then
+export LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu:$LD_LIBRARY_PATH
+fi
+
 
 echo "::group::Run unit tests ('$1' $CC / $CXX)"
 cd "$GITHUB_WORKSPACE/core"